requirements.txt
pydantic>=2.6
langchain
langgraph
langchain-google-genai
google-generativeai
streamlit
python-dotenv
tavily-python